requirements.txt
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
Click==7.0
dicttoxml==1.7.4
Flask==1.0.2
Flask-Cors==3.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycrypto==2.6.1
PyJWT==1.6.4
pylint==2.1.1
PyYAML==3.13
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.12
typed-ast==1.1.0
urllib3==1.24
Werkzeug==0.14.1
wrapt==1.10.11
xmltodict==0.11.0
